chore: Split up the provider file
It was getting huge!
