+++ internal/provider/app_test.go
@@ -1,0 +1,65 @@
+package provider_test
+
+import (
+	"testing"
+
+	"github.com/coder/terraform-provider-coder/internal/provider"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
+	"github.com/stretchr/testify/require"
+)
+
+func TestApp(t *testing.T) {
+	t.Parallel()
+	resource.Test(t, resource.TestCase{
+		Providers: map[string]*schema.Provider{
+			"coder": provider.New(),
+		},
+		IsUnitTest: true,
+		Steps: []resource.TestStep{{
+			Config: `
+				provider "coder" {
+				}
+				resource "coder_agent" "dev" {
+					os = "linux"
+					arch = "amd64"
+				}
+				resource "coder_app" "code-server" {
+					agent_id = coder_agent.dev.id
+					name = "code-server"
+					icon = "builtin:vim"
+					relative_path = true
+					url = "http://localhost:13337"
+					healthcheck {
+						url = "http://localhost:13337/healthz"
+						interval = 5
+						threshold = 6
+					}
+				}
+				`,
+			Check: func(state *terraform.State) error {
+				require.Len(t, state.Modules, 1)
+				require.Len(t, state.Modules[0].Resources, 2)
+				resource := state.Modules[0].Resources["coder_app.code-server"]
+				require.NotNil(t, resource)
+				for _, key := range []string{
+					"agent_id",
+					"name",
+					"icon",
+					"relative_path",
+					"url",
+					"healthcheck.0.url",
+					"healthcheck.0.interval",
+					"healthcheck.0.threshold",
+				} {
+					value := resource.Primary.Attributes[key]
+					t.Logf("%q = %q", key, value)
+					require.NotNil(t, value)
+					require.Greater(t, len(value), 0)
+				}
+				return nil
+			},
+		}},
+	})
+}
